fix add initial input rate limit for spark streaming backpressure bugs

+++ streaming/src/main/scala/org/apache/spark/streaming/receiver/RateLimiter.scala
@@ -37,7 +37,7 @@ private[receiver] abstract class RateLimiter(conf: SparkConf) extends Logging {
 
   // treated as an upper limit
   private val maxRateLimit = conf.getLong("spark.streaming.receiver.maxRate", Long.MaxValue)
-  private lazy val rateLimiter = GuavaRateLimiter.create(getInitialRateLimit())
+  private lazy val rateLimiter = GuavaRateLimiter.create(getInitialRateLimit().toDouble)
 
   def waitToPush() {
     rateLimiter.acquire()
